Create Server Status.lua
ONLY WORKS WITH PROTOSMASHER INJECTED. This will print the server's FPS, so pretty much how well it's doing. 57-60 it's having an okay time, 50-56 it's getting a little worse and 0-49 most likely dying. I made this script for Kohl's Admin House to tell if someone crashed the server or not. And it's a good way to tell if you need to server hop.